% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_hc_imd_bar.R
\name{chart_hc_imd_bar}
\alias{chart_hc_imd_bar}
\title{chart_hc_imd_bar function}
\usage{
chart_hc_imd_bar(
  df,
  imd_col,
  value_col,
  metric_def,
  highlight_core20 = FALSE,
  show_lbl = FALSE,
  value_dp = 0,
  value_prefix = "",
  order_of_magnitude = ""
)
}
\arguments{
\item{df}{A data frame containing data to plot}

\item{imd_col}{String defining name of column within the data that holds the IMD classification (values should be 1 to 10)}

\item{value_col}{String defining name of column within the data that holds the values to be plotted}

\item{metric_def}{String for the metric being reported (will be used as y-axis title and in tooltip)}

\item{highlight_core20}{Should Core20 be highlighted
\itemize{
\item Default = TRUE
\item TRUE:- the Core20 deciles (IMD = 1 or 2) will be highlighted within the output
\item FALSE:- no emphasis will be applied to highlight any of the deciles
}}

\item{show_lbl}{Should data labels be shown
\itemize{
\item Default = TRUE
\item TRUE:- data labels will be included above each decile bar
\item FALSE:- no data labels will be included in the output (values will still be in tooltip)
}}

\item{value_dp}{No. of decimal places to format values to
\itemize{
\item Default = 0
\item Only values between 0 and 3 are valid
\item Decimal values beyond 3 points are not easy to digest and alternative reporting should be considered
}}

\item{value_prefix}{Any value prefix that should be displayed
\itemize{
\item Default = ""
\item Prefix will be included in the axis title and prior to data values (e.g. £1)
}}

\item{order_of_magnitude}{Character to identify if figures should be adjusted to specific abbreviation
\itemize{
\item Default = ""
\item k:- values will be divided by 1,000 and figures will include "k" suffix
\item m:- values will be divided by 1,000,000 and figures will include "m" suffix
\item bn:- values will be divided by 1,000,000,000 and figures will include "bn" suffix
\item tn:- values will be divided by 1,000,000,000,000 and figures will include "tn" suffix
}}
}
\value{
Highchart object for bar chart showing data by IMD decile
}
\description{
Visualisation creation function: Create highchart object showing bar plot IMD decile distribution.

The chart will be a highchart object for a single series chart with a bar per IMD decile. The parameters
supplied will allow basic formatting to be applied to the chart object.

Some chart elements (e.g. title, caption) have been excluded from the function although this can be
added to the highcharts object returned from this function.
}
\examples{
# Example code for chart_hc_imd_bar

# create the test data frame
df_imd_data <- data.frame(
  IMD = c(1,2,3,4,5,6,7,8,9,10),
  VALUE = sample(100000:200000, size = 10 , replace = TRUE)
)

# create chart object
chart_hc_imd_bar(
  df = df_imd_data,
  imd_col = "IMD",
  value_col = "VALUE",
  metric_def = "Example cost values",
  highlight_core20 = TRUE,
  show_lbl = FALSE,
  value_dp = 1,
  value_prefix = "£",
  order_of_magnitude = "k"
)
}
