% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_with_parallelism.R
\name{compute_with_parallelism}
\alias{compute_with_parallelism}
\title{#' Compute (a-la dbplyr) a lazy table with specified degree of parallelism. Adds a parallel hint to all SELECT statements: \verb{SELECT /* +PARALLEL(N) */...} The hint is usually taken into account, but as it is only a hint, the database optimizer ultimately makes a final decision as to which degree of parallelism to use, if any. This is based on things like the query itself and the amount of concurrent use.}
\usage{
compute_with_parallelism(
  lazy_tbl,
  create_table_name,
  n,
  overwrite = F,
  materialize = F
)
}
\arguments{
\item{lazy_tbl}{name of the dbplyr lazy table}

\item{create_table_name}{name of user created table in SQL database}

\item{n}{the degree of parallelism to enforce}

\item{overwrite}{if \code{TRUE}, function will check if table exists, and if it does, will drop it; if \code{FALSE} will throw an error only if the table already exists.}
}
\description{
If hints are already present in any queries, e.g. by including them manually in parsing SQL using \code{dbplyr::sql()}, this function will overwrite them.
}
\examples{
\dontrun{
compute_with_parallelism(table_db, "INT646_TABLE_DB", 32)
}
}
