% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_nhsbsa_fabric.R
\name{con_nhsbsa_fabric}
\alias{con_nhsbsa_fabric}
\title{Connect to a NHSBSA Fabric Lakehouse}
\usage{
con_nhsbsa_fabric(
  driver = "ODBC Driver 18 for SQL Server",
  database,
  username = Sys.getenv("FABRIC_USERNAME"),
  password = Sys.getenv("FABRIC_PASSWORD")
)
}
\arguments{
\item{driver}{String, ODBC driver to connect to the DB. Default is correct
but an argument is provided in case this ever changes.}

\item{database}{String, connection string to the Lakehouse SQL endpoint.}

\item{username}{String, username for DB. Default value is
\code{Sys.getenv("FABRIC_USERNAME")} and we recommend setting
this in an \code{.Renviron} file.}

\item{password}{String, username for DB. Default value is
\code{Sys.getenv("FABRIC_PASSWORD")} and we recommend setting
this in an \code{.Renviron} file.}
}
\value{
Connection to DB. Note: When finished close your connection using
\code{DBI::dbDisconnect()}
}
\description{
Connect to a NHSBSA Fabric Lakehouse. Hopefully you will already have
the ODBC driver installed, but if not you may have to raise a service
request with IT.
}
\examples{
# Initialise connection assuming you have a `.Renviron` file with
# DB_DALP_USERNAME and DB_DALP_PASSWORD set (recommended)
\dontrun{
  con <- nhsbsaR::con_nhsbsa_fabric(database = "Example/Lakehouse/Connection")
}

# Initialise connection without `.Renviron` file (Note: Never store passwords
# in your code)
\dontrun{
  con <- nhsbsaR::con_nhsbsa_fabric(
    database = "Example/Lakehouse/Connection",
    username = rstudioapi::showPrompt(),
    password = rstudioapi::askForPassword()
  )
}


# Remember to disconnect your session
\dontrun{
  DBI::dbDisconnect(con)
}


}
