% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_nhsbsa.R
\name{is_color}
\alias{is_color}
\title{Check input colour is correct built-in R recognisable
color names or HEX format}
\usage{
is_color(x)
}
\arguments{
\item{x}{String type of colour palettes}
}
\value{
boolean
}
\description{
Check input colour is correct built-in R recognisable
color names or HEX format
}
\examples{
is_color(c("red", "#FFFFFF"))
}
