% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_nhsbsa.R
\name{is_nhs_color}
\alias{is_nhs_color}
\title{Check input color is pre-defiend NHS colours}
\usage{
is_nhs_color(x)
}
\arguments{
\item{x}{String type of NHSRtheme colour palettes}
}
\value{
boolean
}
\description{
Check input color is pre-defiend NHS colours
}
\examples{
is_nhs_color(c("DarkBlue","Blue","BrightBlue"))
}
