% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle_create_table.R
\name{oracle_create_table}
\alias{oracle_create_table}
\title{Create table from \code{{dbplyr}} chain}
\usage{
oracle_create_table(df, schema_name = NULL, table_name, overwrite = TRUE)
}
\arguments{
\item{df}{Lazy Oracle table}

\item{schema_name}{Name of the schema. Default is NULL (your own).}

\item{table_name}{Name of the table}

\item{overwrite}{Whether to overwrite an existing table}
}
\description{
Pipe into the end of a \code{{dbplyr}} to create a table in your database schema.
}
\examples{
\dontrun{
dplyr::tbl(src = con, from = "DUAL") \%>\%
  oracle_create_table("<schema_name>", "<table_name>")
}

}
