% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle_merge_strings.R
\name{oracle_merge_strings}
\alias{oracle_merge_strings}
\title{Merge Two Strings Together Whilst Retaining an Order of Some Kind}
\usage{
oracle_merge_strings(df, first_col, second_col, merge_col)
}
\arguments{
\item{df}{Lazy Oracle table}

\item{first_col}{First column}

\item{second_col}{Second column}

\item{merge_col}{Name to give merged column}
}
\value{
original df with additional merged column added
}
\description{
This is intended to be used with differing address strings
The result is a 'master' address string that contains all parts of each address
}
\details{
Requires \code{nhsbsaR::oracle_unnest_tokens()}.
}
\examples{
table_db \%>\% nhsbsaR::oracle_merge_strings(ADDRESS_ONE, ADDRESS_TWO)
}
