% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle_unnest_tokens.R
\name{oracle_unnest_tokens}
\alias{oracle_unnest_tokens}
\title{Split a column of an Oracle table into tokens}
\usage{
oracle_unnest_tokens(df, col, drop = TRUE, pattern = "[:space:]")
}
\arguments{
\item{df}{Lazy Oracle table}

\item{col}{Column to be tokenised}

\item{drop}{Whether original input column should get dropped}

\item{regex}{regex to tokenise by}
}
\description{
A rip off of \code{tidytext::unnest_tokens} that will work on lazy Oracle tables.
This function will retain all columns and add TOKEN_NUMBER and TOKEN columns
to the table.
}
\details{
Inspiration:
https://stackoverflow.com/questions/59537458/how-to-pipe-sql-into-rs-dplyr
}
\examples{
\dontrun{
table_db \%>\% nhsbsaR::oracle_unnest_tokens(col = "DUMMY")
}

}
