% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_nhsbsa.R
\name{pool_nhsbsa}
\alias{pool_nhsbsa}
\title{Pool to NHSBSA DB}
\usage{
pool_nhsbsa(
  driver = "Oracle in OraClient19Home1",
  database,
  username = Sys.getenv("DB_USERNAME"),
  password = Sys.getenv("DB_PASSWORD")
)
}
\arguments{
\item{driver}{String, ODBC driver to connect to the DB. Default is correct
but an argument is provided in case this ever changes.}

\item{database}{String, name of the DB. Default value is "DALP.WORLD".}

\item{username}{String, username for DB. Default value is
\code{Sys.getenv("DB_USERNAME")} and we recommend setting this in an
\code{.Renviron} file.}

\item{password}{String, username for DB. Default value is
\code{Sys.getenv("DB_PASSWORD")} and we recommend setting this in an
\code{.Renviron} file.}

\item{server}{String, server address of the DB. Default value is correct but
an argument is provided in case this ever changes.}

\item{port}{Numeric, port of the DB. Default value is correct but an argument
is provided in case this ever changes.}
}
\value{
Pool to DB. Note: When finished close your connection using
\code{pool::poolClose()}
}
\description{
Pool to NHSBSA DB from the Bastion Host. This is particularly useful in
shiny apps but can be used as a general alternative to \code{con_nhsbsa}.
Hopefully you will already have the ODBC driver installed, but if not you may
have to raise a service request with IT.
}
\examples{
# Initialise pool assuming you have a `.Renviron` file with DB_USERNAME
# and DB_PASSWORD set (recommended)
pool <- nhsbsR::pool_nhsbsa()
pool <- nhsbsR::pool_nhsbsa(database = "DWCP")

# Initialise pool without `.Renviron` file (Note: Never store passwords in
# your code)
pool <- nhsbsR::pool_nhsbsa(
  username = rstudioapi::showPrompt(),
  password = rstudioapi::askForPassword()
)

# Boring DBI methods...
DBI::dbGetQuery(conn = pool, statement = "SELECT * FROM dual")
DBI::dbReadTable(conn = pool, name = "DUAL") # Case sensitive in ODBC

# Cool dbplyr methods... (see https://dbplyr.tidyverse.org)
dplyr::tbl(src = pool, from = "DUAL") # Case sensitive in ODBC
dplyr::tbl(src = pool, from = dbplyr::in_schema(schema = "SYS", table = "DUAL") # Case sensitive in ODBC
dplyr::tbl(src = pool, from = dbplyr::sql("SELECT * FROM dual"))

# Remember to disconnect your session
pool::poolClose(pool)

}
