% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_nhsbsa.R
\name{query_nhsbsa}
\alias{query_nhsbsa}
\title{Query NHSBSA DBs and return results}
\usage{
query_nhsbsa(con, query)
}
\arguments{
\item{con}{a DB connection. Use \code{con_nhsbsa()}}

\item{query}{a file path to a .sql file or character string containing a SQL
query}
}
\value{
a dataframe object
}
\description{
Use a DB connection built with \code{con_nhsbsa()} to deploy a query to the
database and return its results.
}
\examples{
# build connection to DB. Requirse environment variables to be set
con <- nhsbsaR::con_nhsbsa(database = "DWCP")

# create SQL file with query
file.create("example.sql")
writeLines(con = "example.sql", "SELECT * FROM all_tab_cols")

df1 <- query_nhsbsa(con = con, query = "example.sql")

# use query string
df2 <- query_nhsbsa(con = con, query = "select * from all_tab_cols")

}
