% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sql.R
\name{run_sql}
\alias{run_sql}
\title{Execute SQL code on the database}
\usage{
run_sql(script, conn = con, return_output = F)
}
\arguments{
\item{script}{path to the SQL script}

\item{conn}{connection object created using \code{nhsbsaR::con_nhsbsa()}}

\item{return_output}{set to TRUE if the script is expected to return an output from a SELECT statement.}
}
\description{
Execute native SQL code saved in the script on the database. Will execute one statement per script. If the script contains more than one statement, will execute the first statement (e.g. if the script contains \verb{drop table xxx; create table xxx;}, only the DROP TABLE statement will be executed). Handles the presence of SQL variable declarations and comments (in \verb{/* */} and \verb{--}).
}
\examples{
\dontrun{
con <- nhsbsaR::con_nhsbsa(database = "DALP")

# E.g. a create table statement
run_sql("INTXXX_BASE_TABLE.sql")

# E.g. a select statement
df <- run_sql("INTXXX_SUMMARY_DATA.sql")
}
}
