% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_nhsbsa.R
\name{scale_colour_nhsbsa}
\alias{scale_colour_nhsbsa}
\title{Scale colour an NHSBSA ggplot}
\usage{
scale_colour_nhsbsa(palette = NA, n_discrete = 0, reverse = FALSE, ...)
}
\arguments{
\item{palette, }{String type of colour palette. Default is NA, otherwise
should be one one \code{c("gender", "gradient", "highlight")}.}

\item{n_discrete, }{Integer number of discrete categories (if any). Default
is 0.}

\item{reverse}{Boolean, to reverse the palette. Default is FALSE.}

\item{..., }{arguments to pass on}
}
\value{
colour palette
}
\description{
Scale colour an NHSBSA ggplot using the NHSBSA colour palette.
}
\examples{
ggplot2::ggplot(data = ggplot2::diamonds) +
  ggplot2::geom_point(
    mapping = ggplot2::aes(x = carat, y = price, col = carat)
  ) +
  nhsbsaR::scale_colour_nhsbsa(palette = "gradient")
}
