% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_nhsbsa.R
\name{scale_fill_nhsbsa}
\alias{scale_fill_nhsbsa}
\title{Scale fill an NHSBSA ggplot}
\usage{
scale_fill_nhsbsa(palette = NA, n_discrete = 0, reverse = FALSE, ...)
}
\arguments{
\item{palette, }{String type of colour palette. Default is NA, otherwise
should be one one \code{c("gender", "gradient", "highlight")}.}

\item{n_discrete, }{Integer number of discrete categories (if any). Default
is 0.}

\item{reverse}{Boolean, to reverse the palette. Default is FALSE.}

\item{..., }{arguments to pass on}
}
\value{
Colour palette
}
\description{
Scale fill an NHSBSA ggplot using the NHSBSA colour palette.
}
\examples{
ggplot2::ggplot(data = ggplot2::diamonds) +
  ggplot2::geom_bar(mapping = ggplot2::aes(x = cut, fill = cut)) +
  nhsbsaR::scale_fill_nhsbsa()
ggplot2::ggplot(data = ggplot2::diamonds) +
  ggplot2::geom_bar(mapping = ggplot2::aes(x = cut, fill = cut)) +
  nhsbsaR::scale_fill_nhsbsa(palette = "gradient", n_discrete = 5)
}
