% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nhsbsa.R
\name{theme_nhsbsa_gg}
\alias{theme_nhsbsa_gg}
\title{Theme an NHSBSA ggplot}
\usage{
theme_nhsbsa_gg(plot, family = "sans")
}
\arguments{
\item{family, }{String font family to use (this must be installed and loaded).
Default value is "sans" (e.g. "TT Arial").}
}
\value{
ggplot
}
\description{
Theme an NHSBSA ggplot. We would like to take several steps to brand a
standard ggplot to an NHSBSA looking chart. These include:
\itemize{
\item Make the font of any text "sans" or another \code{family} parameter.
\item Colour any text NHS Black
\item Colour background NHS White
\item Make the title and axis titles bold
\item Left align the subtitle
\item Move the legend to the top with no background or key background
\item Remove axis titles (but set them bold incase they are added later). Note
that to add axis titles you must first add the element back in like this:
\code{ggplot2::theme(axis.title.x = ggplot2::element_text()) + ggplot2::xlab()}
\item Colour axis lines NHS Mid Grey
\item Remove axis ticks
\item Remove all gridlines
\item Colour the background of the plot white
\item Colour background NHS White
\item For facets, bold and left align the titles and colour the panel background
}
}
\examples{
# Standard usage
ggplot2::ggplot(data = ggplot2::diamonds) +
  ggplot2::geom_bar(mapping = ggplot2::aes(x = cut)) +
  nhsbsR::theme_nhsbsa_gg()

# Add an axis label
ggplot2::ggplot(data = ggplot2::diamonds) +
  ggplot2::geom_bar(mapping = ggplot2::aes(x = cut)) +
  nhsbsR::theme_nhsbsa_gg() +
  ggplot2::theme(axis.title.x = ggplot2::element_text()) +
  ggplot2::xlab("Cut")

}
