% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nhsbsa.R
\name{theme_nhsbsa_hc}
\alias{theme_nhsbsa_hc}
\title{Custom NHSBSA highcharter theme}
\usage{
theme_nhsbsa_hc(.hc, palette = NA, stack = "normal", family = "TT Arial")
}
\arguments{
\item{palette}{Which colour palette to use from the \code{nhsbsaR} package.}

\item{stack}{Stack option for highcharter.}
}
\value{
Theme data
}
\description{
Based on the nhsbsaR highcharter theme, since it returns a list we can edit
it to the specific theme for this shiny app.
}
