% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nhsbsa.R
\name{theme_nhsbsa_highchart}
\alias{theme_nhsbsa_highchart}
\title{Theme an NHSBSA highcharter plot}
\usage{
theme_nhsbsa_highchart(
  .hc,
  palette = NA,
  stack = "normal",
  family = "TT Arial"
)
}
\arguments{
\item{.hc}{highchart object to be styled}

\item{palette}{Which colour palette to use from the \code{nhsbsaR} package.}

\item{stack}{Stack option for highcharter.}

\item{family, }{String font family to use (this must be installed and loaded).
Default value is "TT Arial"}
}
\value{
highchart object with theme styling applied
}
\description{
Theme an NHSBSA highcharter plot. We would like to take several steps to
brand a standard highcharter plot to an NHSBSA looking chart. These include:
\itemize{
\item Make the font of any text "TT Arial" or another \code{family} parameter.
\item Colour any text NHS Black
\item Colour background NHS White
\item Make the title and axis titles bold
\item Move the legend to the top and set the hover colour to NHS Mid Grey
\item Colour axis ticks and lines NHS Mid Grey
\item Remove yaxis gridlines (x axis has no gridlines by default)
}
}
