% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_nhsbsa.R
\name{palette_nhsbsa}
\alias{palette_nhsbsa}
\title{Generate NHSBSA colour palette}
\usage{
palette_nhsbsa(palette = NA, reverse = FALSE)
}
\arguments{
\item{palette, }{String type of colour palette. Default is NA, otherwise
should be one one \code{c("gender", "gradient", "highlight")}.}

\item{reverse}{Boolean, to reverse the palette. Default is FALSE.}
}
\value{
Colour palette
}
\description{
Generate an accessible and appropriate colour palette for NHSBSA charts that
is inline with the NHS identity. Currently supports up to 8 colours.
}
\examples{
nhsbsR::palette_nhsbsa()
nhsbsR::palette_nhsbsa(palette = "gender")
nhsbsR::palette_nhsbsa(palette = "gradient")
nhsbsR::palette_nhsbsa(palette = "highlight")

}
